+++ src/collective/dms/basecontent/dmsdocument.py
@@ -8,10 +8,11 @@ from plone.dexterity.schema import DexteritySchemaPolicy
 
 from plone.supermodel import model
 from collective.dms.basecontent.relateddocs import RelatedDocs
-from collective.z3cform.rolefield.field import LocalRolesToPrincipals
+from dexterity.localrolesfield.field import LocalRolesField
 
 from . import _
 from .widget import AjaxChosenMultiFieldWidget
+
 
 class IDmsDocument(model.Schema):
     """Schema for DmsDocument"""
@@ -21,18 +22,16 @@ class IDmsDocument(model.Schema):
         required=False,
     )
 
-    treating_groups = LocalRolesToPrincipals(
+    treating_groups = LocalRolesField(
         title=_(u"Treating groups"),
         required=False,
-        roles_to_assign=('Editor',),
         value_type=schema.Choice(vocabulary=u'collective.dms.basecontent.treating_groups',)
     )
     # form.widget(treating_groups=AjaxChosenMultiFieldWidget)
 
-    recipient_groups = LocalRolesToPrincipals(
+    recipient_groups = LocalRolesField(
         title=_(u"Recipient groups"),
         required=False,
-        roles_to_assign=('Reader',),
         value_type=schema.Choice(vocabulary=u'collective.dms.basecontent.recipient_groups')
     )
     form.widget(recipient_groups=AjaxChosenMultiFieldWidget)
@@ -49,10 +48,6 @@ class DmsDocument(Container):
     implements(IDmsDocument)
     # disable local roles inheritance
     __ac_local_roles_block__ = True
-
-    # Needed by collective.z3cform.rolefield
-    treating_groups = FieldProperty(IDmsDocument['treating_groups'])
-    recipient_groups = FieldProperty(IDmsDocument['recipient_groups'])
 
 
 class DmsDocumentSchemaPolicy(DexteritySchemaPolicy):